backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
google-generativeai==0.3.2
openai-whisper==20231117
numpy==1.26.4
python-dotenv==1.0.1
spacy==3.5.0
textstat==0.7.3
language-tool-python==2.7.1
